Move config.py
